t7 init

+++ src/common/utils/io.cpp
@@ -63,7 +63,7 @@ namespace utils::io
 
 			if (size > -1)
 			{
-				data->resize(static_cast<uint32_t>(size));
+				data->resize(size);
 				stream.read(const_cast<char*>(data->data()), size);
 				stream.close();
 				return true;
